requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
qdrant-client==1.11.1
sentence-transformers==3.1.1
pymupdf==1.24.10
google-generativeai==0.8.2
python-dotenv==1.0.1
pydantic==2.9.2
python-docx>=1.1.0
beautifulsoup4>=4.12.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.9